# openapi.json
{
  "components": {
    "schemas": {
...
      "CreateUserRequest": {
        "properties": {
          "email": {
            "type": "string"
          },
          "must_change_password": {
            "type": "boolean"
          },
          "password": {
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ],
        "type": "object"
      },
...
      "CurrentUserResponse": {
        "properties": {
          "email": {
            "type": "string"
          },
          "id": {
            "description": "Stable user id (UUID).",
            "type": "string"
          },
          "must_change_password": {
            "description": "`true` when the user was provisioned with a temporary password and\nmust rotate it before any other action.",
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "email",
          "must_change_password"
        ],
        "type": "object"
      },
...
      "UpdateUserRequest": {
        "properties": {
          "active": {
            "description": "Whether the user is active (can sign in).",
            "type": "boolean"
          },
          "password": {
            "description": "New password. Omit to keep the current one.",
            "nullable": true,
            "type": "string"
          }
        },
        "required": [
          "active"
        ],
        "type": "object"
      },
      "UserResponse": {
        "properties": {
          "active": {
            "type": "boolean"
          },
          "email": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "must_change_password": {
            "type": "boolean"
          },
          "source": {
            "description": "`\"local\"` (password-authenticated) or `\"managed\"` (mlsh-cloud OAuth).",
            "type": "string"
          }
        },
        "required": [
          "id",
          "email",
          "source",
          "active",
          "must_change_password"
        ],
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "API for managing the MLSH overlay network.",
    "license": {
      "identifier": "MIT",
      "name": "MIT"
    },
    "title": "MLSH Control API",
    "version": "1.0.0"
  },
  "openapi": "3.0.3",
  "paths": {
...
    "/api/v1/users": {
      "get": {
        "operationId": "list_users",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/UserResponse"
                  },
                  "type": "array"
                }
              }
            },
            "description": "Users"
          },
          "401": {
            "description": "Not authenticated"
          }
        },
        "summary": "List every user in the cluster.",
        "tags": [
          "Users"
        ]
      },
      "post": {
        "operationId": "create_user",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            },
            "description": "User created"
          },
          "400": {
            "description": "Invalid request"
          },
          "409": {
            "description": "Email already exists, or managed-mode cluster"
          }
        },
        "summary": "Create a local user. Rejected in managed mode (managed clusters provision\nusers from mlsh-cloud OAuth).",
        "tags": [
          "Users"
        ]
      }
    },
    "/api/v1/users/current": {
      "get": {
        "operationId": "get_current",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurrentUserResponse"
                }
              }
            },
            "description": "Current user"
          },
          "401": {
            "description": "Not authenticated"
          }
        },
        "summary": "Get the currently authenticated user.",
        "tags": [
          "Users"
        ]
      }
    },
    "/api/v1/users/{user}": {
      "delete": {
        "operationId": "delete_user",
        "parameters": [
          {
            "description": "User id (UUID)",
            "in": "path",
            "name": "user",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "User deleted"
          },
          "401": {
            "description": "Not authenticated or TOTP step-up failed"
          },
          "403": {
            "description": "Self-delete or last-user-delete refused"
          },
          "404": {
            "description": "User not found"
          }
        },
        "summary": "Delete a user. Refuses self-deletion and last-user deletion. Step-up\nrequired when the caller has verified TOTP.",
        "tags": [
          "Users"
        ]
      },
      "get": {
        "operationId": "get_user",
        "parameters": [
          {
            "description": "User id (UUID)",
            "in": "path",
            "name": "user",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            },
            "description": "User"
          },
          "404": {
            "description": "User not found"
          }
        },
        "summary": "Get a user by id.",
        "tags": [
          "Users"
        ]
      },
      "put": {
        "operationId": "update_user",
        "parameters": [
          {
            "description": "User id (UUID)",
            "in": "path",
            "name": "user",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "User updated"
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Not authenticated or TOTP step-up failed"
          }
        },
        "summary": "Replace a user (active state and optionally password). Step-up required\nwhen the caller has verified TOTP.",
        "tags": [
          "Users"
        ]
      }
    },
...
  }
}